Add QueryCollector
- Move query sorting from ErrorQueue to QueryCollector
